requirements.txt
google-cloud-firestore
google-cloud-aiplatform
google-genai
google-auth-oauthlib
google-auth-httplib2
google-api-python-client
fastapi
uvicorn
python-dotenv
pydantic
httpx
aiofiles
